refactor: auto-split test_api_security.py
